Migrate API server from Spark to Javalin
This was done because Spark is no longer maintained, and does not get things like security updates (not that we really needed those, this was stupid simple). Because of how the Javalin's json serializer works, some methods were not compatible. These were written to move away from Google's Gson, and to just use built in Java arrays with Javalin supports. This has the benefit of this plugin not needing Gson anymore. There is a note in the code that goes into more detail which will be removed in the next commit.